+- coverted double to int because we are using ints for the money

+++ Monopoly/Tiles/Town.cs
@@ -11,10 +11,10 @@ namespace Monopoly
 {
     class Town : Tile, Buildable
     {
-        private double price;
+        private int price;
         private PlayerObject owner;
 
-        public Town(double price) : base(new GetMoneyCommand(400))
+        public Town(int price) : base(new GetMoneyCommand(400))
         {
             this.price = price;
         }
